test: refactor configuration step definitions
Split the definitions of configuration-related steps in two classes:
- `TestEnvironment` contains generic test configuration steps,
  like setting the locale or the location of test files
- `EPUBCheckConfiguration` contains the steps used to conifugre an
  EPUBCheck instance (version, mode, profile, etc)
